requirements.txt
fastapi==0.115.6
sqlalchemy[asyncio]==2.0.36
aiosqlite==0.20.0
openai==1.58.1
python-dotenv==1.0.1
pydantic==2.10.3
